Refactor: moved assets into assets folder
