+++ frontend/src/Components/test/__test__/test.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import test from '../test';
+
+describe('test component', () => {
+  it('Does something...', () => {
+
+  });
+});
