+++ frontend/src/Components/test/test.jsx
@@ -1,0 +1,6 @@
+import './test.css';
+import React from 'react';
+
+export const test = () => <h1>test lal</h1>;
+
+export default test;
