+++ frontend/src/assets/fonts/Klavika/klavika-regular-italic.otf
